// src/client/fluxdocsdRoutes.ts
export interface Fluxdocs {
  path: string
  package: string
  name: string
  kind: string
  headline: string
  fluxType: string
  description?: string
  fluxParameters?: FluxdocsParameters[]
}

export interface FluxdocsParameters {
  description?: string
  headline: string
  name: string
  required: boolean
}